requirements.txt
Flask>=2.3
PyMuPDF>=1.23
python-docx>=0.8.11
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
numpy>=1.24
gunicorn>=20.1.0
